// src/app/modules/cpassapi/model/destinatarioEvasione.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Destinatario } from './destinatario';
import { RigaEvasione } from './rigaEvasione';
import { Settore } from './settore';
import { StatoElOrdine } from './statoElOrdine';
import { TestataEvasione } from './testataEvasione';


export interface DestinatarioEvasione { 
    cap?: string;
    contatto?: string;
    dataCancellazione?: Date;
    dataCreazione?: Date;
    dataModifica?: Date;
    destinatarioOrdine?: Destinatario;
    email?: string;
    id?: string;
    indirizzo?: string;
    localita?: string;
    numCivico?: string;
    optlock?: string;
    progressivo?: number;
    provincia?: string;
    rigaEvasiones?: Array<RigaEvasione>;
    settore?: Settore;
    statoElOrdine?: StatoElOrdine;
    telefono?: string;
    testataEvasione?: TestataEvasione;
    utenteCancellazione?: string;
    utenteCreazione?: string;
    utenteModifica?: string;
}