// src/app/modules/cpassapi/model/testataEvasione.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DestinatarioEvasione } from './destinatarioEvasione';
import { DocumentoTrasporto } from './documentoTrasporto';
import { Fornitore } from './fornitore';
import { Settore } from './settore';
import { Stato } from './stato';
import { TipoEvasione } from './tipoEvasione';
import { Ufficio } from './ufficio';
import { Utente } from './utente';


export interface TestataEvasione { 
    dataAutorizzazione?: Date;
    dataCancellazione?: Date;
    dataConferma?: Date;
    dataConsegna?: Date;
    dataCreazione?: Date;
    dataInserimento?: Date;
    dataInvioContabilita?: Date;
    dataModifica?: Date;
    dataRipartizione?: Date;
    descrizione?: string;
    destinatarioEvasiones?: Array<DestinatarioEvasione>;
    documentoConsegna?: string;
    documentoDataConsegna?: Date;
    documentoTrasporto?: DocumentoTrasporto;
    evasioneAnno?: number;
    evasioneNumero?: number;
    fatturaAnno?: number;
    fatturaCodice?: string;
    fatturaNumero?: string;
    fatturaProtocolloAnno?: number;
    fatturaProtocolloNumero?: number;
    fatturaTipo?: string;
    fatturaTotale?: number;
    fatturaTotaleLiquidabile?: number;
    fornitore?: Fornitore;
    id?: string;
    note?: string;
    optlock?: string;
    settore?: Settore;
    stato?: Stato;
    tipoEvasione?: TipoEvasione;
    totaleConIva?: number;
    ufficio?: Ufficio;
    utenteCancellazione?: string;
    utenteCompilatore?: Utente;
    utenteCreazione?: string;
    utenteModifica?: string;
}