// src/app/modules/cpassapi/model/causaleSospensioneEvasione.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Ente } from './ente';


export interface CausaleSospensioneEvasione { 
    causaleSospensioneCodice?: string;
    causaleSospensioneDescrizione?: string;
    causaleSospensioneId?: number;
    ente?: Ente;
    id?: string;
    utenteCancellazione?: string;
    utenteCreazione?: string;
    utenteModifica?: string;
}