// src/app/modules/cpassapi/model/settoreIndirizzo.ts
/*
* SPDX-FileCopyrightText: Copyright 2019 - 2020 | CSI Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Contabilità passiva
 * API per il backend della suite di contabilità passiva.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Settore } from './settore';


export interface SettoreIndirizzo { 
    id?: string;
    indirizzo?: string;
    numCivico?: string;
    localita?: string;
    provincia?: string;
    cap?: string;
    contatto?: string;
    email?: string;
    telefono?: string;
    settore?: Settore;
}